task 13

+++ 0x02-functions_nested_loops/102-fibonacci.c
@@ -8,14 +8,14 @@ int main(void)
 {
 	int x = 50;
 	int i;
-	unsigned int a = 1, b = 2, next;
+	unsigned long a = 1, b = 2, next;
 
-	printf("%u, %u, ", a, b);
+	printf("%lu, %lu, ", a, b);
 
 	for (i = 3; i <= x; i++)
 	{
 		next = a + b;
-		printf("%u", next);
+		printf("%lu", next);
 
 		if (i < x)
 		{
